// packages/web-app/src/modules/onboarding/models/AntiVirusSoftware.ts
export enum AntiVirusSoftware {
  Adaware = 'Adaware',
  AdvancedSystemCareUltimate = 'Advanced SystemCare Ultimate',
  Avast = 'Avast',
  AVG = 'AVG',
  Avira = 'Avira',
  BitDefender = 'BitDefender',
  BullGuard = 'BullGuard',
  Comodo = 'Comodo',
  EScan = 'e-Scan',
  ESETNOD32 = 'ESET-NOD32',
  FProt = 'F-Prot',
  FSecure = 'F-Secure',
  GData = 'GData',
  K7 = 'K7',
  Kaspersky = 'Kaspersky',
  Malwarebytes = 'Malwarebytes',
  McAfeeSecurity = 'McAfee Security',
  Norton = 'Norton',
  Panda = 'Panda',
  PCMatic = 'PCMatic',
  Qihoo360 = 'Qihoo-360',
  SecureAgeAPEX = 'SecureAge APEX',
  Sophos = 'Sophos',
  TotalAV = 'TotalAV',
  TrendMicro = 'TrendMicro',
  Twister = 'Twister',
  VIPRE = 'VIPRE',
  Webroot = 'Webroot',
  WindowsDefender = 'Windows Defender',
  Zillya = 'Zillya',
}